* Number of improvements and refactorings to wpf example * removed dependency on WpfToolkit, should build fine on teamCity
just some background: This demo will eventually have a number of examples showcasing various elements of Ngenerics in Wpf. The first one is showcasing sorting. For this reason was developed a very light/quick chart control (still not fast enough) which could display thousands of points without flickering. it currently experiments with various rendering methods, in order to find the most optimal one.


git-svn-id: http://ngenerics.googlecode.com/svn/trunk@306 39bbdeee-833e-11de-8ff2-33b0a5e28824